# [WIP] improve graph performance