--- a/package.json
+++ b/package.json
@@ -38,6 +38,5 @@
     "semi": true,
     "singleQuote": true,
     "trailingComma": "es5"
-  },
-  "dependencies": {}
+  }
 }