--- a/packages/foam-vscode/package.json
+++ b/packages/foam-vscode/package.json
@@ -683,7 +683,7 @@
     "@types/dateformat": "^3.0.1",
     "@types/jest": "^29.5.3",
     "@types/lodash": "^4.14.157",
-    "@types/markdown-it": "^12.0.1",
+    "@types/markdown-it": "^12.0.4",
     "@types/micromatch": "^4.0.1",
     "@types/node": "^13.11.0",
     "@types/picomatch": "^2.2.1",