--- a/packages/foam-vscode/package.json
+++ b/packages/foam-vscode/package.json
@@ -712,6 +712,7 @@
     "wait-for-expect": "^3.0.2"
   },
   "dependencies": {
+    "@types/linkify-it": "^3.0.5",
     "dateformat": "4.5.1",
     "detect-newline": "^3.1.0",
     "github-slugger": "^1.4.0",
@@ -720,8 +721,8 @@
     "lodash": "^4.17.21",
     "lru-cache": "^7.14.1",
     "markdown-it-regex": "^0.2.0",
-    "path-browserify": "^1.0.1",
     "mnemonist": "^0.39.8",
+    "path-browserify": "^1.0.1",
     "remark-frontmatter": "^2.0.0",
     "remark-parse": "^8.0.2",
     "remark-wiki-link": "^0.0.4",