@riccardoferretti 2024-09-04
Sharing some high level thoughts here.
As I was doing some research in the approach, I came across the `trie` data structure, which fundamentally feels like what we are kinda implementing here.
As I looked into it, I realized that using the reversed path (similar to what we do in `getShortestIdentifier`) and saving in in a trie would be the cleanest solution, as it would out of the box replace the existing indexing while giving us the features and performance we want.

What do you think? Again, this was at a cursory look, I might be missing something