@pderaaij 2024-09-06
Never really looked into the trie datastructure. Will do some research and fiddle around!