@pderaaij 2024-09-25
@riccardoferretti I've updated the implementation with my WIP so far. 

It has a little impact on performance. Compared to the earlier version, I now load the `10000s` workspace in 2 to 3 seconds. Which still is a massive improvement compared to the current implementation. I do feel the performance impact is acceptable given the ease of maintenance this implementation brings in the code.

I have one remaining issue and I hope you can help. Originally, a Map holds the insertion order. TrieMap obviously does not do that. But, this means that `resources()` needs to return a sorted Iterator. I can't find a correct way to sort the values and then return an `IteratableIterator`. Perhaps you have an idea here?

If we can resolve this issue, all tests will be green so we can do some further testing and enhancing before shipping it.