@riccardoferretti 2024-10-11
@pderaaij I think it's related to the markdown-it version change