@pderaaij 2024-10-11
Well... I found the little bugger:

https://github.com/markdown-it/linkify-it/issues/111