@pderaaij 2024-10-11
I am giving up on this PR. Going to do a fresh branch on master and open a new PR. Not trusting all the mess this has created.